Split db-specific Module traits into more specialized traits for Sql, Renderer and Jdbc
