+++ build.sbt
@@ -151,6 +151,7 @@ lazy val jdbc = project
 lazy val mysql = project
   .in(file("mysql"))
   .dependsOn(jdbc % "compile->compile;test->test")
+  .dependsOn(driver)
   .settings(stdSettings("zio-sql-mysql"))
   .settings(buildInfoSettings("zio.sql.mysql"))
   .settings(
@@ -168,6 +169,7 @@ lazy val mysql = project
 lazy val oracle = project
   .in(file("oracle"))
   .dependsOn(jdbc % "compile->compile;test->test")
+  .dependsOn(driver)
   .settings(stdSettings("zio-sql-oracle"))
   .settings(buildInfoSettings("zio.sql.oracle"))
   .settings(
@@ -185,6 +187,7 @@ lazy val oracle = project
 lazy val postgres = project
   .in(file("postgres"))
   .dependsOn(jdbc % "compile->compile;test->test")
+  .dependsOn(driver)
   .settings(stdSettings("zio-sql-postgres"))
   .settings(buildInfoSettings("zio.sql.postgres"))
   .settings(
@@ -202,6 +205,7 @@ lazy val postgres = project
 lazy val sqlserver = project
   .in(file("sqlserver"))
   .dependsOn(jdbc % "compile->compile;test->test")
+  .dependsOn(driver)
   .settings(stdSettings("zio-sql-sqlserver"))
   .settings(buildInfoSettings("zio.sql.sqlserver"))
   .settings(
