+++ mysql/src/main/scala/zio/sql/mysql/MysqlJdbcModule.scala
@@ -1,0 +1,5 @@
+package zio.sql.mysql
+
+import zio.sql.Jdbc
+
+trait MysqlJdbcModule extends MysqlRenderModule with Jdbc
