+++ mysql/src/test/scala/zio/sql/mysql/MysqlRunnableSpec.scala
@@ -5,7 +5,7 @@ import zio.sql.{ ConnectionPoolConfig, JdbcRunnableSpec }
 import zio.test._
 import zio.ZLayer
 
-trait MysqlRunnableSpec extends JdbcRunnableSpec with MysqlModule {
+trait MysqlRunnableSpec extends JdbcRunnableSpec with MysqlJdbcModule {
 
   private def connProperties(user: String, password: String): Properties = {
     val props = new Properties
