+++ oracle/src/main/scala/zio/sql/oracle/OracleJdbcModule.scala
@@ -1,0 +1,5 @@
+package zio.sql.oracle
+
+import zio.sql.Jdbc
+
+trait OracleJdbcModule extends OracleRenderModule with Jdbc
