+++ oracle/src/main/scala/zio/sql/oracle/OracleSqlModule.scala
@@ -1,0 +1,11 @@
+package zio.sql.oracle
+
+import zio.sql.Sql
+
+trait OracleSqlModule extends Sql { self =>
+
+  object OracleFunctionDef {
+    val Sind = FunctionDef[Double, Double](FunctionName("sind"))
+  }
+
+}
