+++ postgres/src/main/scala/zio/sql/postgresql/PostgresJdbcModule.scala
@@ -1,0 +1,5 @@
+package zio.sql.postgresql
+
+import zio.sql.Jdbc
+
+trait PostgresJdbcModule extends PostgresRenderModule with Jdbc
