+++ postgres/src/test/scala/zio/sql/postgresql/PostgresRunnableSpec.scala
@@ -5,7 +5,7 @@ import zio.test._
 import java.util.Properties
 import zio.sql.{ ConnectionPoolConfig, JdbcRunnableSpec }
 
-trait PostgresRunnableSpec extends JdbcRunnableSpec with PostgresModule {
+trait PostgresRunnableSpec extends JdbcRunnableSpec with PostgresJdbcModule {
 
   def autoCommit: Boolean = true
