+++ sqlserver/src/main/scala/zio/sql/sqlserver/SqlServerJdbcModule.scala
@@ -1,0 +1,5 @@
+package zio.sql.sqlserver
+
+import zio.sql.Jdbc
+
+trait SqlServerJdbcModule extends SqlServerRenderModule with Jdbc
